requirements.txt
pyinstaller==4.3
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pillow==8.2.0